Fix all warnings

+++ source/Player.h
@@ -6,8 +6,8 @@ struct Player
     float y;
     float width;
     float height;
-    int turnDirection; // -1 for left, +1 for right
-    int walkDirection; // -1 for back, +1 for front
+    float turnDirection; // -1 for left, +1 for right
+    float walkDirection; // -1 for back, +1 for front
     float rotationAngle;
     float walkSpeed;
     float turnSpeed;
